Client/requirements.txt
websocket-client
keyboard==0.13.5
requests
fastapi
uvicorn[standard]
motor
pydantic
rich
pynput